+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/logging/LogMessageKeys.java
@@ -216,6 +216,7 @@ public enum LogMessageKeys {
     RESOLVER_KEY("resolver_key"),
     RESOLVER_PATH("resolver_path"),
     RESOLVER_METADATA("resolver_metadata"),
+    RESOLVER_VALUE("resolver_value"),
     CACHED_KEY("cached_key"),
     CACHED_STATE("cached_state"),
     READ_STATE("read_state"),