+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/keyspace/ExtendedDirectoryLayer.java
@@ -156,6 +156,11 @@ public class ExtendedDirectoryLayer extends LocatableResolver {
     }
 
     @Override
+    protected CompletableFuture<Void> putReverse(@Nonnull final FDBRecordContext context, final long value, @Nonnull final String key) {
+        return database.getReverseDirectoryCache().putOrReplaceForTesting(context, wrap(key), value);
+    }
+
+    @Override
     public CompletableFuture<Void> setMapping(FDBRecordContext context, String key, ResolverResult value) {
         return read(context, key).thenCombine(readInReverseCacheSubpspace(context.getTimer(), value.getValue()), (maybeRead, maybeReverseRead) -> {
             maybeRead.ifPresent(read -> {